backend/python-outreach/requirements.txt
telethon==1.36.0
supabase==2.6.0
python-dotenv==1.0.1
loguru==0.7.2
asyncio
aiohttp

